requirements.txt
Pillow==9.2.0
streamlit==1.12.0
streamlit_timeline==0.0.2
extra_streamlit_components==0.1.56